move plugins code to service
