+++ Sources/GeneralKit/Services/PluginService/PluginService.swift
@@ -1,0 +1,13 @@
+//
+//  Copyright © 2021 Rosberry. All rights reserved.
+//
+
+import ArgumentParser
+
+public protocol HasPluginService {
+    var pluginService: PluginService { get }
+}
+
+public protocol PluginService {
+    func main(command: ParsableCommand.Type) throws
+}
